IDT with bug

+++ kernel/runtime/itoa.h
@@ -1,9 +1,11 @@
 #ifndef _ITOA_H_
 #define _ITOA_H_
 
+#include <runtime/types.h>
+
 extern "C" {
 
-char* itoa(int num, char* str, int base);
+char* itoa(u64 num, char* str, int base);
 
 }
